+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Context.java
@@ -31,6 +31,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import java.util.function.Predicate;
 
 import org.graalvm.polyglot.impl.AbstractPolyglotImpl.AbstractContextImpl;
 
@@ -245,6 +246,7 @@ public final class Context implements AutoCloseable {
         private InputStream in;
         private Map<String, String> options;
         private Map<String, String[]> arguments;
+        private Predicate<String> classFilter;
 
         Builder(String... onlyLanguages) {
             Objects.requireNonNull(onlyLanguages);
@@ -321,10 +323,26 @@ public final class Context implements AutoCloseable {
         }
 
         /**
+         * Sets a class filter that allows to limit the classes that are allowed to be loaded by
+         * guest languages. If the filter returns <code>true</code> then the class is accessible,
+         * else it is not accessible and throws an guest language error when accessed.
+         *
+         * @param classFilter a predicate that returns <code>true</code> or <code>false</code> for a
+         *            java qualified class name.
+         *
+         * @since 1.0
+         */
+        public Builder setJavaClassFilter(Predicate<String> classFilter) {
+            Objects.requireNonNull(classFilter);
+            this.classFilter = classFilter;
+            return this;
+        }
+
+        /**
          * Set an option for this language {@link Context context}. If one of the set option keys or
          * values is invalid then an {@link IllegalArgumentException} is thrown when the context is
          * {@link #build() built}. The given key and value must not be <code>null</code>. Options
-         * for the engine or instruments must be specified using the
+         * for the engine or instruments can be specified using the
          * {@link Engine.Builder#setOption(String, String) engine builder}.
          *
          * @see Language#getOptions() To list all available options for a {@link Language language}.
@@ -432,16 +450,15 @@ public final class Context implements AutoCloseable {
                 }
                 engineBuilder.setBoundEngine(true);
                 engine = engineBuilder.build();
-                return engine.impl.createContext(null, null, null, Collections.emptyMap(),
+                return engine.impl.createContext(null, null, null, classFilter, Collections.emptyMap(),
                                 arguments == null ? Collections.emptyMap() : arguments,
                                 onlyLanguages);
             } else {
-                return engine.impl.createContext(out, err, in,
+                return engine.impl.createContext(out, err, in, classFilter,
                                 options == null ? Collections.emptyMap() : options,
                                 arguments == null ? Collections.emptyMap() : arguments,
                                 onlyLanguages);
             }
-
         }
 
     }